+++ lib/cacheflow/redis.rb
@@ -1,6 +1,5 @@
 module Cacheflow
   module Redis
-    # redis 5 / redis-client
     module ClientNotifications
       def call(command, redis_config)
         payload = {
@@ -12,18 +11,6 @@ module Cacheflow
       end
 
       def call_pipelined(commands, redis_config)
-        payload = {
-          commands: commands
-        }
-        ActiveSupport::Notifications.instrument("query.redis", payload) do
-          super
-        end
-      end
-    end
-
-    # redis 4
-    module Notifications
-      def logging(commands)
         payload = {
           commands: commands
         }
@@ -50,11 +37,5 @@ module Cacheflow
   end
 end
 
-if defined?(RedisClient)
-  RedisClient.register(Cacheflow::Redis::ClientNotifications)
-elsif defined?(Redis)
-  # redis < 5
-  Redis::Client.prepend(Cacheflow::Redis::Notifications)
-end
-
+RedisClient.register(Cacheflow::Redis::ClientNotifications)
 Cacheflow::Redis::Instrumenter.attach_to(:redis)